keep local branche update

+++ hack/verify-godeps.sh
@@ -28,33 +28,50 @@ preload-dep() {
   org="$1"
   project="$2"
   sha="$3"
+  # project_dir ($4) is optional, if unset we will generate it
+  if [[ -z ${4:-} ]]; then
+    project_dir="${GOPATH}/src/${org}/${project}.git"
+  else
+    project_dir="${4}"
+  fi
 
-  org_dir="${GOPATH}/src/${org}"
-  mkdir -p "${org_dir}"
-  pushd "${org_dir}" > /dev/null
-    git clone "https://${org}/${project}.git" > /dev/null 2>&1
-    pushd "${org_dir}/${project}" > /dev/null
-      git checkout "${sha}" > /dev/null 2>&1
-    popd > /dev/null
+  echo "**HACK** preloading dep for ${org} ${project} at ${sha} into ${project_dir}"
+  git clone "https://${org}/${project}" "${project_dir}" > /dev/null 2>&1
+  pushd "${project_dir}" > /dev/null
+    git checkout "${sha}"
   popd > /dev/null
 }
 
 KUBE_ROOT=$(dirname "${BASH_SOURCE}")/..
 source "${KUBE_ROOT}/hack/lib/init.sh"
-kube::golang::verify_godep_version
 
 readonly branch=${1:-${KUBE_VERIFY_GIT_BRANCH:-master}}
 if ! [[ ${KUBE_FORCE_VERIFY_CHECKS:-} =~ ^[yY]$ ]] && \
   ! kube::util::has_changes_against_upstream_branch "${branch}" 'Godeps/' && \
-  ! kube::util::has_changes_against_upstream_branch "${branch}" 'vendor/'; then
+  ! kube::util::has_changes_against_upstream_branch "${branch}" 'vendor/' && \
+  ! kube::util::has_changes_against_upstream_branch "${branch}" 'hack/.*godep'; then
   exit 0
 fi
 
-# Create a nice clean place to put our new godeps
-_tmpdir="$(mktemp -d -t gopath.XXXXXX)"
+if [[ -z ${TMP_GOPATH:-} ]]; then
+  # Create a nice clean place to put our new godeps
+  _tmpdir="$(mktemp -d -t gopath.XXXXXX)"
+else
+  # reuse what we might have saved previously
+  _tmpdir="${TMP_GOPATH}"
+fi
+
+if [[ -z ${KEEP_TMP:-} ]]; then
+    KEEP_TMP=false
+fi
 function cleanup {
-  echo "Removing ${_tmpdir}"
-  rm -rf "${_tmpdir}"
+  if [ "${KEEP_TMP}" == "true" ]; then
+    echo "Leaving ${_tmpdir} for you to examine or copy. Please delete it manually when finished. (rm -rf ${_tmpdir})"
+  else
+    echo "Removing ${_tmpdir}"
+    rm -rf "${_tmpdir}"
+  fi
+  export GODEP=""
 }
 trap cleanup EXIT
 
@@ -67,41 +84,63 @@ _kubetmp="${_kubetmp}/kubernetes"
 
 # Do all our work in the new GOPATH
 export GOPATH="${_tmpdir}"
-cd "${_kubetmp}"
 
-# Build the godep tool
-go get -u github.com/tools/godep 2>/dev/null
-GODEP="${GOPATH}/bin/godep"
-pin-godep() {
-  pushd "${GOPATH}/src/github.com/tools/godep" > /dev/null
-    git checkout "$1"
-    "${GODEP}" go install
-  popd > /dev/null
-}
-# Use to following if we ever need to pin godep to a specific version again
-#pin-godep 'v63'
+pushd "${_kubetmp}" 2>&1 > /dev/null
+  # Restore the Godeps into our temp directory
+  hack/godep-restore.sh
 
-# Fill out that nice clean place with the kube godeps
-echo "Starting to download all kubernetes godeps. This takes a while"
-"${GODEP}" restore
-echo "Download finished"
+  # Destroy deps in the copy of the kube tree
+  rm -rf ./Godeps ./vendor
 
-# Destroy deps in the copy of the kube tree
-rm -rf ./Godeps ./vendor
+  # For some reason the kube tree needs to be a git repo for the godep tool to
+  # run. Doesn't make sense.
+  git init > /dev/null 2>&1
 
-# For some reason the kube tree needs to be a git repo for the godep tool to
-# run. Doesn't make sense.
-git init > /dev/null 2>&1
+  # Recreate the Godeps using the nice clean set we just downloaded
+  hack/godep-save.sh
+popd 2>&1 > /dev/null
 
-# Recreate the Godeps using the nice clean set we just downloaded
-hack/godep-save.sh
+ret=0
 
-# Test for diffs
-if ! _out="$(diff -Naupr --ignore-matching-lines='^\s*\"GoVersion\":' --ignore-matching-line='^\s*\"GodepVersion\":' --ignore-matching-lines='^\s*\"Comment\":' ${KUBE_ROOT}/Godeps/Godeps.json ${_kubetmp}/Godeps/Godeps.json)"; then
-  echo "Your Godeps.json is different:"
-  echo "${_out}"
-  echo "Godeps Verify failed."
-  exit 1
+pushd "${KUBE_ROOT}" 2>&1 > /dev/null
+  # Test for diffs
+  if ! _out="$(diff -Naupr --ignore-matching-lines='^\s*\"GoVersion\":' --ignore-matching-line='^\s*\"GodepVersion\":' --ignore-matching-lines='^\s*\"Comment\":' Godeps/Godeps.json ${_kubetmp}/Godeps/Godeps.json)"; then
+    echo "Your Godeps.json is different:"
+    echo "${_out}"
+    echo "Godeps Verify failed."
+    echo "${_out}" > godepdiff.patch
+    echo "If you're seeing this locally, run the below command to fix your Godeps.json:"
+    echo "patch -p0 < godepdiff.patch"
+    echo "(The above output can be saved as godepdiff.patch if you're not running this locally)"
+    echo "(The patch file should also be exported as a build artifact if run through CI)"
+    KEEP_TMP=true
+    if [[ -f godepdiff.patch && -d "${ARTIFACTS_DIR:-}" ]]; then
+      echo "Copying patch to artifacts.."
+      cp godepdiff.patch "${ARTIFACTS_DIR:-}/"
+    fi
+    ret=1
+  fi
+
+  if ! _out="$(diff -Naupr -x "BUILD" -x "AUTHORS*" -x "CONTRIBUTORS*" vendor ${_kubetmp}/vendor)"; then
+    echo "Your vendored results are different:"
+    echo "${_out}"
+    echo "Godeps Verify failed."
+    echo "${_out}" > vendordiff.patch
+    echo "If you're seeing this locally, run the below command to fix your directories:"
+    echo "patch -p0 < vendordiff.patch"
+    echo "(The above output can be saved as godepdiff.patch if you're not running this locally)"
+    echo "(The patch file should also be exported as a build artifact if run through CI)"
+    KEEP_TMP=true
+    if [[ -f vendordiff.patch && -d "${ARTIFACTS_DIR:-}" ]]; then
+      echo "Copying patch to artifacts.."
+      cp vendordiff.patch "${ARTIFACTS_DIR:-}/"
+    fi
+    ret=1
+  fi
+popd 2>&1 > /dev/null
+
+if [[ ${ret} > 0 ]]; then
+  exit ${ret}
 fi
 
 echo "Godeps Verified."
